Maximum Subarray return the subarray

+++ Maximum_subarray.java
@@ -3,7 +3,7 @@ public class Maximum_subarray{
     public static void main(String [] args){
         int [] arr = {-2, -3, 4, -1, -2, 1, 5, -3};
 
-        long  maxi = Long.MIN_VALUE;
+        int maxi = Integer.MIN_VALUE;
         int sum = 0;
 
         for(int i = 0; i < arr.length; i++){
